Finish authentication

+++ app/Http/Controllers/MeetingController.php
@@ -5,13 +5,15 @@ namespace App\Http\Controllers;
 use App\Meeting;
 use Illuminate\Http\Request;
 use Carbon\Carbon;
-
+use JWTAuth;
 class MeetingController extends Controller
 {
 
     public function __construct()
     {
-        //$this->middleware('name');
+        $this->middleware('jwt.auth', ['only' => [
+            'upate', 'store', 'destroy'
+        ]]);
     }
 
     /**
@@ -52,14 +54,18 @@ class MeetingController extends Controller
         $this->validate($request, [
             'title' => 'required',
             'description' => 'required',
-            'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required',
+            'time' => 'required|date_format:YmdHie'
         ]);
+
+        if(!$user = JWTAuth::parseToken()->authenticate())
+        {
+            return response()->json(['msg' => 'User not found'], 404);
+        }
 
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
+        $user_id = $user->id;
 
         $meeting = new Meeting([
             'time' => Carbon::createFromFormat('YmdHie', $time),
@@ -125,14 +131,18 @@ class MeetingController extends Controller
         $this->validate($request, [
             'title' => 'required',
             'description' => 'required',
-            'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required',
+            'time' => 'required|date_format:YmdHie'
         ]);
+
+        if(!$user = JWTAuth::parseToken()->authenticate())
+        {
+            return response()->json(['msg' => 'User not found'], 404);
+        }
 
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
+        $user_id = $user->id;
 
         $meeting = [
             'title' => $title,
@@ -181,6 +191,16 @@ class MeetingController extends Controller
     public function destroy($id)
     {
         $meeting = Meeting::findOrFail($id);
+
+        if(!$user = JWTAuth::parseToken()->authenticate())
+        {
+            return response()->json(['msg' => 'User not found'], 404);
+        }
+
+        if(!$meeting->users()->where('users.id', $user->id)->first()){
+            return response()->json(['msg' => 'user not registered for meeting, update not successful'], 401);
+        }
+
         $users = $meeting->users;
         $meeting->users()->detach();
 
@@ -190,7 +210,7 @@ class MeetingController extends Controller
             }
             return response()->json(['msg' => 'deletion failed'], 404);
         }
-        
+
         $response = [
             'msg' => 'Meeting deleted',
             'create' => [
